change routes and table css
