+++ src/components/SideBar.jsx
@@ -9,7 +9,7 @@ export default class SideBar extends React.Component {
   render () {
     return (
       <Menu  noOverlay>
-        <img src ="https://cdn3.iconfinder.com/data/icons/medcare/512/Stethoscope-256.png" align="middle" alt="Icone"/>
+        <img src ="https://cdn3.iconfinder.com/data/icons/medcare/512/Stethoscope-256.png" align="middle" alt="Icone" />
         <li><a href="/DoctorForm"><span className="sub_icon glyphicon glyphicon-plus"></span>  Cadastrar médico</a></li>
         <li><a href="/ScheduleTable"><span className="sub_icon glyphicon glyphicon-calendar "></span>  Visualizar tabela</a></li>
         <li><a  href="/DoctorStatus"><span className="sub_icon glyphicon glyphicon-sort "></span>  Alterar status</a></li>
