+++ src/css/NavBar.css
@@ -6,7 +6,7 @@
 
 .navbar{
   width: 100%;
-  font-size: 15px;
+  font-size: 10px;
   list-style-type: none;
   border: none;
 }
