+++ src/css/ScheduleTable.css
@@ -1,4 +1,4 @@
-/* h1 {
+h1 {
   margin-top: 62px;
   text-align: center;
   line-height: 70px;
@@ -25,7 +25,7 @@ h3 {
 }
 .espace{
   margin-bottom: 5em;
-} */
+}
 
 
 
@@ -108,6 +108,11 @@ h3 {
     text-align: center;
     text-shadow: 0 1px 0 rgba(255,255,255,0.2);
     border-radius: 15px 15px 15px 0px;
+}
+
+.change-color{
+  background: linear-gradient(#007095,#20C3D3);
+  border-radius: 10px;
 }
 
 .form-style-5 input[type="submit"],
